удаление юзера

+++ src/main/java/com/cwrsoi/model/UserDtls.java
@@ -10,8 +10,7 @@ public class UserDtls {
 
     @Id
     @GeneratedValue//(strategy = GenerationType.IDENTITY)
-    //private int id;
-    private Long id;
+    private int id;
     @Column(name="first_name")
     private String firstName;
     @Column(name="last_name")
